Initial successfully built solution, VS2019

+++ Project/Syser/Source/NtstatStr.h
@@ -4,7 +4,7 @@ typedef struct _NTSTATUS_STR
 {
 	char* Description;
 	char* Name;
-	int Code;
+	unsigned long Code;
 }NTSTATUS_STR,*P_NTSTATUS_STR;
 extern NTSTATUS_STR gNtstatusTable[];
 #endif //_NTSTATSTR_H_
